layout/base/tests/chrome/dialog_with_positioning_window.xul

<?xml version="1.0" encoding="UTF-8" ?>
<?xml-stylesheet href="chrome://global/skin/" type="text/css"?>
<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="setTimeout(runTest, 0)">
  <vbox>
    <text value="powered by example.com" style="padding: 16px;"/>
  </vbox>
  <hbox id="t" style="position: fixed; right: 16px; bottom: 16px;">
    <button label="OK"/>
  </hbox>
<script><![CDATA[
var SimpleTest = window.opener.wrappedJSObject.SimpleTest;
var SpecialPowers = window.opener.wrappedJSObject.SpecialPowers;
var is = window.opener.wrappedJSObject.is;
var ok = window.opener.wrappedJSObject.ok;

// We run this off a setTimeout from onload, because the XUL window
// only does its intrinsic-height layout after the load event has
// finished
function runTest() {
  var t = document.getElementById("t");
  var tBottom = t.getBoundingClientRect().bottom;
  is(tBottom, document.documentElement.getBoundingClientRect().bottom - 16,
     "check fixed-pos element t bottom positioned correctly");
  ok(tBottom < 200, "fixed-pos element t bottom must be sane, less than 200 (got " + tBottom + ")");
  window.close();
  SimpleTest.finish();
}
]]></script>
</window>
